+++ modules/002-deckhouse/hooks/pkg/apis/validation/validation_webhooks.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package apis
+package validation
 
 import (
 	log "github.com/sirupsen/logrus"
